+++ src/generator/annotations.ts
@@ -34,3 +34,4 @@ export const DTO_RELATION_MODIFIERS_ON_UPDATE = [
 ];
 export const DTO_TYPE_FULL_UPDATE = /@DtoTypeFullUpdate/;
 export const DTO_CAST_TYPE = /@DtoCastType/;
+export const CUSTOM_VALIDATOR = /@CustomValidator/;