+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - output NestJs pluralized model folders.
 - add `@ApiProperty`.type for `@DtoCastType'
 - support nested parentheses for validators, e.g.: @Matches(/^[a-z\d](?:[a-z\d]|-(?=[a-z\d])){0,38}$/i)
+- add annotation `@CustomValidator`
 
 ## [1.17.4] - 2023-03-04